Refactor JWT key handling by consolidating functions into common module and removing deprecated keys_id module
